@@ -1,7 +1,7 @@
 {
   "name": "quickapp",
-  "version": "4.0.0",
-  "description": "ASP.NET Core 3.1/Angular 9 startup project template with complete login, user and role management. Plus other useful services for Quick Application Development",
+  "version": "5.0.0",
+  "description": "ASP.NET Core 5.0/Angular 11 startup project template with complete login, user and role management. Plus other useful services for Quick Application Development",
   "author": {
     "name": "Ebenezer Monney",
     "email": "info@ebenmonney.com",
@@ -22,53 +22,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.4",
-    "@angular/common": "~9.1.4",
-    "@angular/compiler": "~9.1.4",
-    "@angular/core": "~9.1.4",
-    "@angular/forms": "~9.1.4",
-    "@angular/platform-browser": "~9.1.4",
-    "@angular/platform-browser-dynamic": "~9.1.4",
-    "@angular/router": "~9.1.4",
-    "@ng-select/ng-select": "^4.0.0",
-    "@ngx-translate/core": "^12.1.2",
-    "@swimlane/ngx-datatable": "^17.0.0",
-    "angular-oauth2-oidc": "^9.2.1",
-    "bootstrap": "^4.4.1",
+    "@angular/animations": "~11.2.9",
+    "@angular/common": "~11.2.9",
+    "@angular/compiler": "~11.2.9",
+    "@angular/core": "~11.2.9",
+    "@angular/forms": "~11.2.9",
+    "@angular/platform-browser": "~11.2.9",
+    "@angular/platform-browser-dynamic": "~11.2.9",
+    "@angular/router": "~11.2.9",
+    "@ng-select/ng-select": "^6.1.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@swimlane/ngx-datatable": "^19.0.0",
+    "angular-oauth2-oidc": "^10.0.3",
+    "bootstrap": "^4.6.0",
     "bootstrap-toggle": "^2.2.2",
-    "bootswatch": "^4.4.1",
-    "chart.js": "^2.9.3",
+    "bootswatch": "^4.6.0",
+    "chart.js": "^2.9.4",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.5.0",
-    "ng2-charts": "^2.3.0",
-    "ngx-bootstrap": "^5.6.1",
-    "ngx-toasta": "^1.0.3",
+    "jquery": "^3.6.0",
+    "ng2-charts": "^2.4.2",
+    "ngx-bootstrap": "^6.2.0",
+    "ngx-toasta": "^2.0.0",
     "popper.js": "^1.16.1",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.4",
-    "@angular/cli": "~9.1.4",
-    "@angular/compiler-cli": "~9.1.4",
-    "@angular/language-service": "~9.1.4",
+    "@angular-devkit/build-angular": "~0.1102.8",
+    "@angular/cli": "~11.2.8",
+    "@angular/compiler-cli": "~11.2.9",
+    "@types/jasmine": "~3.6.0",
+    "@types/jquery": "^3.5.5",
     "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/jquery": "^3.3.36",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~5.0.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.1.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~3.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "npm-run-all": "^4.1.5",
-    "protractor": "~5.4.3",
+    "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.8.3"
+    "typescript": "~4.1.5"
   }
 }
