@@ -19,48 +19,47 @@ import { AuthGuard } from './services/auth-guard.service';
 import { Utilities } from './services/utilities';
 
 
-
 @Injectable()
 export class LowerCaseUrlSerializer extends DefaultUrlSerializer {
-    parse(url: string): UrlTree {
-        const possibleSeparators = /[?;#]/;
-        const indexOfSeparator = url.search(possibleSeparators);
-        let processedUrl: string;
+  parse(url: string): UrlTree {
+    const possibleSeparators = /[?;#]/;
+    const indexOfSeparator = url.search(possibleSeparators);
+    let processedUrl: string;
 
-        if (indexOfSeparator > -1) {
-            const separator = url.charAt(indexOfSeparator);
-            const urlParts = Utilities.splitInTwo(url, separator);
-            urlParts.firstPart = urlParts.firstPart.toLowerCase();
+    if (indexOfSeparator > -1) {
+      const separator = url.charAt(indexOfSeparator);
+      const urlParts = Utilities.splitInTwo(url, separator);
+      urlParts.firstPart = urlParts.firstPart.toLowerCase();
 
-            processedUrl = urlParts.firstPart + separator + urlParts.secondPart;
-        } else {
-            processedUrl = url.toLowerCase();
-        }
-
-        return super.parse(processedUrl);
+      processedUrl = urlParts.firstPart + separator + urlParts.secondPart;
+    } else {
+      processedUrl = url.toLowerCase();
     }
+
+    return super.parse(processedUrl);
+  }
 }
 
 
 const routes: Routes = [
-    { path: '', component: HomeComponent, canActivate: [AuthGuard], data: { title: 'Home' } },
-    { path: 'login', component: LoginComponent, data: { title: 'Login' } },
-    { path: 'customers', component: CustomersComponent, canActivate: [AuthGuard], data: { title: 'Customers' } },
-    { path: 'products', component: ProductsComponent, canActivate: [AuthGuard], data: { title: 'Products' } },
-    { path: 'orders', component: OrdersComponent, canActivate: [AuthGuard], data: { title: 'Orders' } },
-    { path: 'settings', component: SettingsComponent, canActivate: [AuthGuard], data: { title: 'Settings' } },
-    { path: 'about', component: AboutComponent, data: { title: 'About Us' } },
-    { path: 'home', redirectTo: '/', pathMatch: 'full' },
-    { path: '**', component: NotFoundComponent, data: { title: 'Page Not Found' } }
+  { path: '', component: HomeComponent, canActivate: [AuthGuard], data: { title: 'Home' } },
+  { path: 'login', component: LoginComponent, data: { title: 'Login' } },
+  { path: 'customers', component: CustomersComponent, canActivate: [AuthGuard], data: { title: 'Customers' } },
+  { path: 'products', component: ProductsComponent, canActivate: [AuthGuard], data: { title: 'Products' } },
+  { path: 'orders', component: OrdersComponent, canActivate: [AuthGuard], data: { title: 'Orders' } },
+  { path: 'settings', component: SettingsComponent, canActivate: [AuthGuard], data: { title: 'Settings' } },
+  { path: 'about', component: AboutComponent, data: { title: 'About Us' } },
+  { path: 'home', redirectTo: '/', pathMatch: 'full' },
+  { path: '**', component: NotFoundComponent, data: { title: 'Page Not Found' } }
 ];
 
 
 @NgModule({
-    imports: [RouterModule.forRoot(routes)],
-    exports: [RouterModule],
-    providers: [
-        AuthService,
-        AuthGuard,
-        { provide: UrlSerializer, useClass: LowerCaseUrlSerializer }]
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule],
+  providers: [
+    AuthService,
+    AuthGuard,
+    { provide: UrlSerializer, useClass: LowerCaseUrlSerializer }]
 })
 export class AppRoutingModule { }
