@@ -3,7 +3,7 @@
 // www.ebenmonney.com/templates
 // =============================
 
-import { TestBed, async } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
@@ -34,63 +34,63 @@ import { AccountService } from '../services/account.service';
 import { AccountEndpoint } from '../services/account-endpoint.service';
 
 describe('AppComponent', () => {
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            imports: [
-                HttpClientModule,
-                FormsModule,
-                RouterTestingModule,
-                TranslateModule.forRoot({
-                    loader: {
-                        provide: TranslateLoader,
-                        useClass: TranslateLanguageLoader
-                    }
-                }),
-                NgxDatatableModule,
-                OAuthModule.forRoot(),
-                ToastaModule.forRoot(),
-                TooltipModule.forRoot(),
-                PopoverModule.forRoot(),
-                ModalModule.forRoot()
-            ],
-            declarations: [
-                AppComponent,
-                LoginComponent,
-                NotificationsViewerComponent
-            ],
-            providers: [
-                AuthService,
-                AlertService,
-                ConfigurationService,
-                ThemeManager,
-                AppTitleService,
-                AppTranslationService,
-                NotificationService,
-                NotificationEndpoint,
-                AccountService,
-                AccountEndpoint,
-                LocalStoreManager,
-                OidcHelperService
-            ]
-        }).compileComponents();
-    }));
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        HttpClientModule,
+        FormsModule,
+        RouterTestingModule,
+        TranslateModule.forRoot({
+          loader: {
+            provide: TranslateLoader,
+            useClass: TranslateLanguageLoader
+          }
+        }),
+        NgxDatatableModule,
+        OAuthModule.forRoot(),
+        ToastaModule.forRoot(),
+        TooltipModule.forRoot(),
+        PopoverModule.forRoot(),
+        ModalModule.forRoot()
+      ],
+      declarations: [
+        AppComponent,
+        LoginComponent,
+        NotificationsViewerComponent
+      ],
+      providers: [
+        AuthService,
+        AlertService,
+        ConfigurationService,
+        ThemeManager,
+        AppTitleService,
+        AppTranslationService,
+        NotificationService,
+        NotificationEndpoint,
+        AccountService,
+        AccountEndpoint,
+        LocalStoreManager,
+        OidcHelperService
+      ]
+    }).compileComponents();
+  });
 
-    it('should create the app', () => {
-        const fixture = TestBed.createComponent(AppComponent);
-        const app = fixture.componentInstance;
-        expect(app).toBeTruthy();
-    });
+  it('should create the app', () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app).toBeTruthy();
+  });
 
-    it(`should have as title 'QuickApp'`, () => {
-        const fixture = TestBed.createComponent(AppComponent);
-        const app = fixture.componentInstance as AppComponent;
-        expect(app.appTitle).toEqual('QuickApp');
-    });
+  it(`should have as title 'QuickApp'`, () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance as AppComponent;
+    expect(app.appTitle).toEqual('QuickApp');
+  });
 
-    it('should render Loaded! in a h1 tag', () => {
-        const fixture = TestBed.createComponent(AppComponent);
-        fixture.detectChanges();
-        const compiled = fixture.nativeElement;
-        expect(compiled.querySelector('h1').textContent).toContain('Loaded!');
-    });
+  it('should render Loaded! in a h1 tag', () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    fixture.detectChanges();
+    const compiled = fixture.nativeElement;
+    expect(compiled.querySelector('h1').textContent).toContain('Loaded!');
+  });
 });
