@@ -92,7 +92,7 @@ export class ConfigurationService {
     get showDashboardBanner() {
         return this._showDashboardBanner != null ? this._showDashboardBanner : ConfigurationService.defaultShowDashboardBanner;
     }
-    public static readonly appVersion: string = '4.0.0';
+    public static readonly appVersion: string = '5.0.0';
 
     // ***Specify default configurations here***
     public static readonly defaultLanguage: string = 'en';
